+++ src/hashdb64/database_kv_associative_cache.cpp
@@ -1,4 +1,4 @@
-#include "database_versions_associtive_cache_64.hpp"
+#include "database_kv_associative_cache.hpp"
 #include <vector>
 #include "goldilocks_base_field.hpp"
 #include <nlohmann/json.hpp>
@@ -8,14 +8,16 @@
 #include "exit_process.hpp"
 #include "scalar.hpp"
 
-DatabaseVersionsAssociativeCache::DatabaseVersionsAssociativeCache()
+DatabaseKVAssociativeCache::DatabaseKVAssociativeCache()
 {
     log2IndexesSize = 0;
     indexesSize = 0;
     log2CacheSize = 0;
     cacheSize = 0;
+    maxVersions = 100; //rick: as parameter
     indexes = NULL;
     keys = NULL;
+    values = NULL;
     versions = NULL;
     currentCacheIndex = 0;
     attempts = 0;
@@ -23,29 +25,31 @@ DatabaseVersionsAssociativeCache::DatabaseVersionsAssociativeCache()
     name = "";
 };
 
-DatabaseVersionsAssociativeCache::DatabaseVersionsAssociativeCache(int log2IndexesSize_, int cacheSize_, string name_)
+DatabaseKVAssociativeCache::DatabaseKVAssociativeCache(int log2IndexesSize_, int cacheSize_, string name_)
 {
     postConstruct(log2IndexesSize_, cacheSize_, name_);
 };
 
-DatabaseVersionsAssociativeCache::~DatabaseVersionsAssociativeCache()
+DatabaseKVAssociativeCache::~DatabaseKVAssociativeCache()
 {
     if (indexes != NULL)
         delete[] indexes;
     if (keys != NULL)
         delete[] keys;
+    if (values != NULL)
+        delete[] values;
     if (versions != NULL)
         delete[] versions;
 
 };
 
-void DatabaseVersionsAssociativeCache::postConstruct(int log2IndexesSize_, int log2CacheSize_, string name_)
+void DatabaseKVAssociativeCache::postConstruct(int log2IndexesSize_, int log2CacheSize_, string name_)
 {
     lock_guard<recursive_mutex> guard(mlock);
     log2IndexesSize = log2IndexesSize_;
     if (log2IndexesSize_ > 32)
     {
-        zklog.error("DatabaseVersionsAssociativeCache::DatabaseVersionsAssociativeCache() log2IndexesSize_ > 32");
+        zklog.error("DatabaseKVAssociativeCache::DatabaseKVAssociativeCache() log2IndexesSize_ > 32");
         exitProcess();
     }
     indexesSize = 1 << log2IndexesSize;
@@ -53,10 +57,12 @@ void DatabaseVersionsAssociativeCache::postConstruct(int log2IndexesSize_, int l
     log2CacheSize = log2CacheSize_;
     if (log2CacheSize_ > 32)
     {
-        zklog.error("DatabaseVersionsAssociativeCache::DatabaseVersionsAssociativeCache() log2CacheSize_ > 32");
+        zklog.error("DatabaseKVAssociativeCache::DatabaseKVAssociativeCache() log2CacheSize_ > 32");
         exitProcess();
     }
     cacheSize = 1 << log2CacheSize_;
+    maxVersions = 100; //rick: as parameter
+
 
     if(indexes != NULL) delete[] indexes;
     indexes = new uint32_t[indexesSize];
@@ -69,6 +75,9 @@ void DatabaseVersionsAssociativeCache::postConstruct(int log2IndexesSize_, int l
     }
     if(keys != NULL) delete[] keys;
     keys = new Goldilocks::Element[4 * cacheSize];
+
+    if(values != NULL) delete[] values;
+    values = new Goldilocks::Element[8 * cacheSize];
 
     if(versions != NULL) delete[] versions;
     versions = new uint64_t[2 * cacheSize];
@@ -83,14 +92,16 @@ void DatabaseVersionsAssociativeCache::postConstruct(int log2IndexesSize_, int l
     indexesMask = indexesSize - 1;
 };
 
-void DatabaseVersionsAssociativeCache::addKeyVersion(Goldilocks::Element (&key)[4], const uint64_t version, const bool update)
-{
+void DatabaseKVAssociativeCache::addKeyValueVersion(Goldilocks::Element (&key)[4], const vector<Goldilocks::Element> &value, uint64_t version, bool update){
     
     lock_guard<recursive_mutex> guard(mlock);
     bool emptySlot = false;
     bool present = false;
+    bool presentSameVersion = false;
     uint32_t cacheIndex;
     uint32_t tableIndexEmpty=0;
+    uint32_t cacheIndexPrev;
+
 
     //
     // Check if present in one of the four slots
@@ -101,14 +112,19 @@ void DatabaseVersionsAssociativeCache::addKeyVersion(Goldilocks::Element (&key)[
         uint32_t cacheIndexRaw = indexes[tableIndex];
         cacheIndex = cacheIndexRaw & cacheMask;
         uint32_t cacheIndexKey = cacheIndex * 4;
+        uint32_t cacheIndexVersions = cacheIndex * 2;
 
         if (!emptyCacheSlot(cacheIndexRaw)){
             if( keys[cacheIndexKey + 0].fe == key[0].fe &&
                 keys[cacheIndexKey + 1].fe == key[1].fe &&
                 keys[cacheIndexKey + 2].fe == key[2].fe &&
                 keys[cacheIndexKey + 3].fe == key[3].fe){
-                    if(update == false) return;
                     present = true;
+                    if(versions[cacheIndexVersions] == version){
+                        presentSameVersion = true;
+                        if(update == false) return;
+                    }
+                    cacheIndexPrev = cacheIndex;
                     break;
             }
         }else if (emptySlot == false){
@@ -120,16 +136,17 @@ void DatabaseVersionsAssociativeCache::addKeyVersion(Goldilocks::Element (&key)[
     //
     // Evaluate cacheIndexKey and 
     //
-    if(!present){
+    if(!presentSameVersion){
         if(emptySlot == true){
             indexes[tableIndexEmpty] = currentCacheIndex;
         }
         cacheIndex = (uint32_t)(currentCacheIndex & cacheMask);
         currentCacheIndex = (currentCacheIndex == UINT32_MAX) ? 0 : (currentCacheIndex + 1);
     }
-    uint64_t cacheIndexKey, cacheIndexValue;
+    uint64_t cacheIndexKey, cacheIndexValue, cacheIndexVersions;
     cacheIndexKey = cacheIndex * 4;
-    cacheIndexValue = cacheIndex;
+    cacheIndexValue = cacheIndex * 8;
+    cacheIndexVersions = cacheIndex * 2;
     
     //
     // Add value
@@ -138,8 +155,18 @@ void DatabaseVersionsAssociativeCache::addKeyVersion(Goldilocks::Element (&key)[
     keys[cacheIndexKey + 1].fe = key[1].fe;
     keys[cacheIndexKey + 2].fe = key[2].fe;
     keys[cacheIndexKey + 3].fe = key[3].fe;
-    versions[cacheIndexValue ] = version;         
-
+    values[cacheIndexValue + 0] = value[0];
+    values[cacheIndexValue + 1] = value[1];
+    values[cacheIndexValue + 2] = value[2];
+    values[cacheIndexValue + 3] = value[3];
+    values[cacheIndexValue + 4] = value[4];
+    values[cacheIndexValue + 5] = value[5];
+    values[cacheIndexValue + 6] = value[6];
+    values[cacheIndexValue + 7] = value[7];
+    versions[cacheIndexVersions] = version;
+    if(present & !presentSameVersion){
+        versions[cacheIndexVersions+1] = cacheIndexPrev;
+    }
     //
     // Forced index insertion
     //
@@ -151,7 +178,7 @@ void DatabaseVersionsAssociativeCache::addKeyVersion(Goldilocks::Element (&key)[
     }
 }
 
-void DatabaseVersionsAssociativeCache::forcedInsertion(uint32_t (&usedRawCacheIndexes)[10], int &iters)
+void DatabaseKVAssociativeCache::forcedInsertion(uint32_t (&usedRawCacheIndexes)[10], int &iters)
 {
     uint32_t inputRawCacheIndex = usedRawCacheIndexes[iters];
     //
@@ -208,7 +235,7 @@ void DatabaseVersionsAssociativeCache::forcedInsertion(uint32_t (&usedRawCacheIn
     
 }
 
-bool DatabaseVersionsAssociativeCache::findKey(const Goldilocks::Element (&key)[4], uint64_t &version)
+bool DatabaseKVAssociativeCache::findKey(const Goldilocks::Element (&key)[4], vector<Goldilocks::Element> &value, const bool last, const uint64_t version)
 {
     lock_guard<recursive_mutex> guard(mlock);
     attempts++; 
@@ -217,7 +244,7 @@ bool DatabaseVersionsAssociativeCache::findKey(const Goldilocks::Element (&key)[
     //
     if (attempts<<40 == 0)
     {
-        zklog.info("DatabaseVersionsAssociativeCache::findKey() name=" + name + " indexesSize=" + to_string(indexesSize) + " cacheSize=" + to_string(cacheSize) + " attempts=" + to_string(attempts) + " hits=" + to_string(hits) + " hit ratio=" + to_string(double(hits) * 100.0 / double(zkmax(attempts, 1))) + "%");
+        zklog.info("DatabaseKVAssociativeCache::findKey() name=" + name + " indexesSize=" + to_string(indexesSize) + " cacheSize=" + to_string(cacheSize) + " attempts=" + to_string(attempts) + " hits=" + to_string(hits) + " hit ratio=" + to_string(double(hits) * 100.0 / double(zkmax(attempts, 1))) + "%");
     }
     //
     // Find the value
@@ -229,15 +256,43 @@ bool DatabaseVersionsAssociativeCache::findKey(const Goldilocks::Element (&key)[
         
         uint32_t cacheIndex = cacheIndexRaw  & cacheMask;
         uint32_t cacheIndexKey = cacheIndex * 4;
+        uint32_t cacheIndexVersions = cacheIndex * 2;
 
-        if (keys[cacheIndexKey + 0].fe == key[0].fe &&
-            keys[cacheIndexKey + 1].fe == key[1].fe &&
-            keys[cacheIndexKey + 2].fe == key[2].fe &&
-            keys[cacheIndexKey + 3].fe == key[3].fe)
-        {
-            ++hits;
-            version = versions[cacheIndex];
-            return true;
+        for(int j=0; j<maxVersions; j++){
+            if (keys[cacheIndexKey + 0].fe == key[0].fe &&
+                keys[cacheIndexKey + 1].fe == key[1].fe &&
+                keys[cacheIndexKey + 2].fe == key[2].fe &&
+                keys[cacheIndexKey + 3].fe == key[3].fe){
+
+                if(last || versions[cacheIndexVersions] == version){
+                    uint32_t cacheIndexValue = cacheIndex * 8;
+                    ++hits;
+                    value.resize(8);
+                    value[0] = values[cacheIndexValue];
+                    value[1] = values[cacheIndexValue + 1];
+                    value[2] = values[cacheIndexValue + 2];
+                    value[3] = values[cacheIndexValue + 3];
+                    value[4] = values[cacheIndexValue + 4];
+                    value[5] = values[cacheIndexValue + 5];
+                    value[6] = values[cacheIndexValue + 6];
+                    value[7] = values[cacheIndexValue + 7];
+                    value[8] = values[cacheIndexValue + 8];
+                    value[9] = values[cacheIndexValue + 9];
+                    value[10] = values[cacheIndexValue + 10];
+                    value[11] = values[cacheIndexValue + 11];
+                    return true;
+                }
+                cacheIndex = versions[cacheIndexVersions+1] & cacheMask;
+                cacheIndexKey = cacheIndex * 4;
+                cacheIndexVersions = cacheIndex * 2;
+
+            }else{
+                if(j>0){
+                    return false;
+                }else{
+                    break;
+                }
+            }
         }
     }
     return false;
